Little improvements
