+++ src/main/java/de/melanx/recipeprinter/renderers/vanilla/BlastingRender.java
@@ -1,4 +1,4 @@
-package de.melanx.recipeprinter.renderers;
+package de.melanx.recipeprinter.renderers.vanilla;
 
 import com.mojang.blaze3d.matrix.MatrixStack;
 import de.melanx.recipeprinter.IRecipeRender;
@@ -6,22 +6,22 @@ import de.melanx.recipeprinter.util.RenderHelper;
 import net.minecraft.client.renderer.IRenderTypeBuffer;
 import net.minecraft.item.ItemStack;
 import net.minecraft.item.Items;
-import net.minecraft.item.crafting.CampfireCookingRecipe;
+import net.minecraft.item.crafting.BlastingRecipe;
 import net.minecraft.item.crafting.IRecipeType;
 
 import javax.annotation.Nullable;
 
-public class CampfireRender implements IRecipeRender<CampfireCookingRecipe> {
+public class BlastingRender implements IRecipeRender<BlastingRecipe> {
 
     @Override
-    public Class<CampfireCookingRecipe> getRecipeClass() {
-        return CampfireCookingRecipe.class;
+    public Class<BlastingRecipe> getRecipeClass() {
+        return BlastingRecipe.class;
     }
 
     @Nullable
     @Override
-    public IRecipeType<? super CampfireCookingRecipe> getRecipeType() {
-        return IRecipeType.CAMPFIRE_COOKING;
+    public IRecipeType<? super BlastingRecipe> getRecipeType() {
+        return IRecipeType.BLASTING;
     }
 
     @Override
@@ -35,8 +35,8 @@ public class CampfireRender implements IRecipeRender<CampfireCookingRecipe> {
     }
 
     @Override
-    public void render(CampfireCookingRecipe recipe, MatrixStack matrixStack, IRenderTypeBuffer buffer) {
+    public void render(BlastingRecipe recipe, MatrixStack matrixStack, IRenderTypeBuffer buffer) {
         SmeltingRender.render(recipe, matrixStack, buffer);
-        RenderHelper.renderItem(matrixStack, buffer, new ItemStack(Items.CAMPFIRE), 5, 41);
+        RenderHelper.renderItem(matrixStack, buffer, new ItemStack(Items.BLAST_FURNACE), 5, 41);
     }
 }
